patch

+++ src/users/users.service.ts
@@ -22,7 +22,9 @@ export class UsersService {
     }
 
     async delete(id: string){
-        await this.userModel.deleteOne({_id :id});
+        if (id !== ""){
+            await this.userModel.deleteOne({_id :id});
+        }
     }
 
     findOne(id:string) {
